# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Graft REST API Sandbox",
    "description": "A session-based sandbox for exploring git repositories using the Graft MCP tools over REST.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://your-app-name.onrender.com",
      "description": "Production Sandbox Server"
    }
  ],
  "paths": {
    "/sessions": {
      "get": {
        "summary": "Create a new Sandbox Session via GET query parameters",
        "description": "Provisions a new isolated session sandbox using URL parameters.",
        "operationId": "createSessionGet",
        "parameters": [
          {
            "name": "repositoryUrl",
            "in": "query",
            "required": false,
            "description": "The URL of the git repository to clone. Must be a publicly accessible repository.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sessionId": {
                      "type": "string",
                      "description": "The unique UUID of the newly created session."
                    }
                  },
                  "required": ["sessionId"]
                }
              }
            }
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "post": {
        "summary": "Create a new Sandbox Session",
        "description": "Provisions a new isolated session sandbox. If a repositoryUrl is provided, it clones that repository into the sandbox.",
        "operationId": "createSession",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "repositoryUrl": {
                    "type": "string",
                    "description": "The URL of the git repository to clone. Must be a publicly accessible repository."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sessionId": {
                      "type": "string",
                      "description": "The unique UUID of the newly created session."
                    }
                  },
                  "required": ["sessionId"]
                }
              }
            }
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
    "/sessions/{sessionId}/tools": {
      "get": {
        "summary": "List Session Tools",
        "description": "Returns a list of all Graft MCP tools available for this specific session, including their JSON schemas.",
        "operationId": "listTools",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "description": "The UUID of the session.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of tools.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tools": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": { "type": "string" },
                          "description": { "type": "string" },
                          "inputSchema": { "type": "object" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{sessionId}/tools/{toolName}": {
      "get": {
        "summary": "Execute a Tool via GET query parameters",
        "description": "Executes a specific Graft tool within the sandbox session using query parameters for arguments.",
        "operationId": "executeToolGet",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "description": "The UUID of the session.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "toolName",
            "in": "path",
            "required": true,
            "description": "The name of the tool to execute (e.g., safe_read, capabilities).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tool executed successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "content": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    },
                    "structuredContent": {
                      "description": "Machine-readable tool output.",
                      "type": "object",
                      "additionalProperties": true
                    },
                    "isError": {
                      "type": "boolean",
                      "default": false
                    },
                    "_meta": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Session or Tool not found."
          }
        }
      },
      "post": {
        "summary": "Execute a Tool",
        "description": "Executes a specific Graft tool within the sandbox session.",
        "operationId": "executeTool",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "description": "The UUID of the session.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "toolName",
            "in": "path",
            "required": true,
            "description": "The name of the tool to execute (e.g., safe_read, capabilities).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "The arguments for the tool, as defined by the tool's schema."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tool executed successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "content": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    },
                    "structuredContent": {
                      "description": "Machine-readable tool output.",
                      "type": "object",
                      "additionalProperties": true
                    },
                    "isError": { 
                      "type": "boolean",
                      "default": false
                    },
                    "_meta": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Session or Tool not found."
          }
        }
      }
    }
  }
}